Move file validation into BscPlugin
