+++ src/bscPlugin/BscPlugin.ts
@@ -1,4 +1,4 @@
-import { isBrsFile } from '../astUtils/reflection';
+import { isBrsFile, isXmlFile } from '../astUtils/reflection';
 import type { BeforeFileTranspileEvent, CompilerPlugin, OnFileValidateEvent, OnGetCodeActionsEvent, ProvideHoverEvent, OnGetSemanticTokensEvent, OnScopeValidateEvent, ProvideCompletionsEvent } from '../interfaces';
 import type { Program } from '../Program';
 import { CodeActionsProcessor } from './codeActions/CodeActionsProcessor';
@@ -8,6 +8,7 @@ import { BrsFileSemanticTokensProcessor } from './semanticTokens/BrsFileSemantic
 import { BrsFilePreTranspileProcessor } from './transpile/BrsFilePreTranspileProcessor';
 import { BrsFileValidator } from './validation/BrsFileValidator';
 import { ScopeValidator } from './validation/ScopeValidator';
+import { XmlFileValidator } from './validation/XmlFileValidator';
 
 export class BscPlugin implements CompilerPlugin {
     public name = 'BscPlugin';
@@ -33,6 +34,8 @@ export class BscPlugin implements CompilerPlugin {
     public onFileValidate(event: OnFileValidateEvent) {
         if (isBrsFile(event.file)) {
             return new BrsFileValidator(event as any).process();
+        } else if (isXmlFile(event.file)) {
+            return new XmlFileValidator(event as any).process();
         }
     }
